filetypes: use aenum.Enum and update documentation
the obj is to have a class representing filetypes, i can inherit from
But Enum are not inheritable. But aenum let be update the original enum.
I wanted to not add deps to the opencf_core, except the magic that have a cmplex code that do what i dont have in mimetypes builtin: read file content to infer type.
The current workaround is quite complex (involding, meta class and abstract) without doing pure inheritance, which lead to a bug while
Now, aenum.Enum dont have support for typing so mypy will recognise.
A workaround i used involve letting loading enum.Enum for mypy and aenum.Enum in compilation
So, now i still dont have actual inheritance but i use only once Enum class the user can modify, which is fine by me
I still have a problem. the enum members i add are not type-hinted in vscode, so not for the users either